添加卡片资源及卡牌信息表格完善 翻牌添加顺序标识

+++ KENDivination/Class/View/KENUiViewChouPai.m
@@ -41,7 +41,7 @@
     [self addSubview:_selectPaiBgView];
     
     UIImage* image = [UIImage imageNamed:@"app_pai_bg.png"];
-    int count = [[KENModel shareModel] getPaiZhenNumber];
+    NSInteger count = [[KENModel shareModel] getPaiZhenNumber];
     float width = 240 / count;
     float space = width;
     if (width < image.size.width) {
